% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_field.R
\name{add_field}
\alias{add_field}
\alias{add_field_type}
\alias{add_checkbox_field}
\alias{add_dropdown_field}
\alias{add_notes_field}
\alias{add_radio_field}
\alias{add_text_field}
\alias{add_truefalse_field}
\title{Add a new field.}
\usage{
add_field(.data, ..., .before = NULL, .after = NULL)

add_field_type(.data, ..., .type)

add_checkbox_field(.data, ...)

add_dropdown_field(.data, ...)

add_notes_field(.data, ...)

add_radio_field(.data, ...)

add_text_field(.data, ...)

add_truefalse_field(.data, ...)
}
\arguments{
\item{.data}{Data frame to append to.}

\item{...}{Name-value pairs, passed on to \code{\link[=tibble]{tibble()}}. Only columns that exist
in \code{.data} can be used, unset columns will get an \code{NA} value.
These arguments are passed on to \code{\link[=tibble]{tibble()}}, and therefore also support
unquote via \code{!!} and unquote-splice via \code{!!!}.}

\item{.before}{One-based row index where to add the new rows,
default: after last row.}

\item{.after}{One-based row index where to add the new rows,
default: after last row.}

\item{.type}{A field type. One of: "checkbox", "dropdown", "notes", "radio",
"text", "truefalse".}
}
\description{
Adds a new field to a tibble containing existing fields.
}
