% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_missing_columns.R
\name{add_missing_columns}
\alias{add_missing_columns}
\title{Add missing columns}
\usage{
add_missing_columns(df)
}
\arguments{
\item{df}{Data frame to which columns should be added.}
}
\description{
Given a data frame containing some columns for a data dictionary, add the
remaining (populated with NAs).
}
\examples{
df <- tibble::tibble(`Variable / Field Name` = "patient_name")
add_missing_columns(df)
}
