% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_choices.R
\name{create_choices}
\alias{create_choices}
\title{Create choices for a radio, dropdown, or checkbox question}
\usage{
create_choices(v)
}
\arguments{
\item{v}{A character vector of choice labels.}
}
\description{
Choices should take the format '1, option1 | 2, option2' etc.
\code{create_choices()} takes a vector and returns a string in the required
format.
}
\examples{
v <- c("red", "green", "blue")
create_choices(v)
}
