% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_template.R
\name{create_dd_template}
\alias{create_dd_template}
\title{Create empty data dictionary template}
\usage{
create_dd_template()
}
\value{
An empty (0-row) data frame with columns corresponding to the column
names required in a REDCap data dictionary.
}
\description{
Create empty data dictionary template
}
