% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_fields.R
\name{create_fields}
\alias{create_fields}
\title{Create one or more fields in a REDCap data dictionary}
\usage{
create_fields(x)
}
\arguments{
\item{x}{An object (list or data frame) containing information for the fields
to be added.}
}
\value{
A tibble containing one or more fields in the data dictionary with
all required columns.
}
\description{
Create one or more fields in a REDCap data dictionary
}
\examples{

x <- list(
  "Variable / Field Name" = "patient_name",
  "Field Label" = "Patient name"
)
create_fields(x)

x <- tibble::tibble(
`Variable / Field Name` = "patient_name",
`Field Label` = "Patient name"
)
create_fields(x)
}
