% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/columns.R
\name{get_dd_cols}
\alias{get_dd_cols}
\title{Get the column names of a REDCap data dictionary}
\usage{
get_dd_cols()
}
\value{
A character vector containing the valid column names.
}
\description{
Provides a character vector of column names. Valid columns in a REDCap data
dictionary are:
\itemize{
\item Variable / Field Name
\item Form Name
\item Section Header
\item Field Type
\item Field Label
\item Choices, Calculations, OR Slider Labels
\item Field Note
\item Text Validation Type OR Show Slider Number
\item Text Validation Min
\item Text Validation Max
\item Identifier?
\item Branching Logic (Show field only if...)
\item Required Field?
\item Custom Alignment
\item Question Number (surveys only)
\item Matrix Group Name
\item Matrix Ranking?
\item Field Annotation
}
}
\examples{
dd_cols <- get_dd_cols()
}
