% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_dd.R
\name{write_dd}
\alias{write_dd}
\title{Export data dictionary as a CSV file}
\usage{
write_dd(df, filename = "data_dictionary.csv")
}
\arguments{
\item{df}{Tibble containing the data dictionary.}

\item{filename}{File name to write to.}
}
\description{
Export data dictionary as a CSV file
}
