% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_fields.R
\name{create_fields}
\alias{create_fields}
\title{Create one or more fields in a REDCap data dictionary}
\usage{
create_fields(..., df = NULL)
}
\arguments{
\item{...}{Elements of the data dictionary to be populated. Can be passed as
individual arguments, or as a list. All columns not described will be
populated with \code{NA}. If passing a list, it must be unquote-spliced (see
examples).}

\item{df}{Optionally, a data frame containing one or more fields.}
}
\value{
A tibble containing one or more fields in the data dictionary.
}
\description{
Create one or more fields in a REDCap data dictionary
}
\examples{
create_fields(
  "Variable / Field Name" = "patient_name",
  "Field Label" = "Patient name"
)

# When passing a list, it must be unquote-spliced
x <- list(
  "Variable / Field Name" = "patient_name",
  "Field Label" = "Patient name"
)
create_fields(!!!x)

x <- list(
`Variable / Field Name` = "patient_name",
`Field Label` = "Patient name"
)
create_fields(!!!x)
}
